Добавил тесты для SocialServices но один не работает из-за уведомлений которые не мокируются. Как их замокать остается загадкой -_-

+++ src/main/java/ru/gwolk/librarysocial/AppBackend/SocialServices/SubscriptionsService.java
@@ -37,9 +37,14 @@ public class SubscriptionsService {
         createAndSetSubsciptionsGrid();
     }
 
-    public void unsubscribe(User subscibedUser) {
+    public void unsubscribe(User subscribedUser) {
         me = getUser();
-        Subscription subscription = subscriptionsRepository.findSubscriptionByUserAndSubscribedUser(me, subscibedUser);
+        Subscription subscription = subscriptionsRepository.findSubscriptionByUserAndSubscribedUser(me, subscribedUser);
+
+        if (subscription == null) {
+            throw new IllegalArgumentException("Подписка не найдена");
+        }
+
         subscriptionsRepository.delete(subscription);
         subscriptionsGrid.setItems(subscriptionsRepository.findAllSubscribedUsers());
     }
